requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
Markdown==3.4.3
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
stripe==9.9.0
